+++ modules/platforms/dotnet/Apache.Ignite.Core.Tests/Apache.Ignite.Core.Tests.csproj
@@ -118,13 +118,13 @@
     <Compile Include="Cache\DataStorageMetricsTest.cs" />
     <Compile Include="Cache\NearCacheTest.cs" />
     <Compile Include="Cache\Platform\PlatformCacheTestCreateDestroy.cs" />
-    <Compile Include="Cache\Platform\CacheNearTopologyChangeTest.cs" />
-    <Compile Include="Cache\Platform\CacheNearTest.cs" />
+    <Compile Include="Cache\Platform\PlatformCacheTopologyChangeTest.cs" />
+    <Compile Include="Cache\Platform\PlatformCacheTest.cs" />
     <Compile Include="Cache\Platform\FailingCacheStore.cs" />
     <Compile Include="Cache\Platform\Foo.cs" />
-    <Compile Include="Cache\Platform\ScanQueryNearCacheFilter.cs" />
-    <Compile Include="Cache\Platform\ScanQueryNoNearCacheFilter.cs" />
-    <Compile Include="Cache\Platform\StoreNoNearCacheFilter.cs" />
+    <Compile Include="Cache\Platform\ScanQueryPlatformCacheFilter.cs" />
+    <Compile Include="Cache\Platform\ScanQueryNoPlatformCacheFilter.cs" />
+    <Compile Include="Cache\Platform\StoreNoPlatformCacheFilter.cs" />
     <Compile Include="Cache\PartitionPreloadTest.cs" />
     <Compile Include="Cache\PersistenceTest.cs" />
     <Compile Include="Cache\PersistentStoreTestObsolete.cs" />
