+++ modules/platforms/dotnet/Apache.Ignite.Core.Tests/Cache/Platform/ScanQueryPlatformCacheFilter.cs
@@ -16,29 +16,44 @@
 
 namespace Apache.Ignite.Core.Tests.Cache.Platform
 {
+    using System.Security;
     using Apache.Ignite.Core.Cache;
     using Apache.Ignite.Core.Resource;
     using NUnit.Framework;
 
     /// <summary>
-    /// Scan query filter that checks whether values come from native near cache.
+    /// Scan query filter that checks whether values come from platform cache.
     /// </summary>
-    public class ScanQueryNoNearCacheFilter : ICacheEntryFilter<int, Foo>
+    public class ScanQueryPlatformCacheFilter : ICacheEntryFilter<int, Foo>
     {
+        /// <summary>
+        /// Gets or sets the cache name.
+        /// </summary>
         public string CacheName { get; set; }
         
+        /// <summary>
+        /// Gets or sets the key that should cause an exception in <see cref="Invoke"/>. 
+        /// </summary>
+        public int? FailKey { get; set; }
+        
+        /// <summary>
+        /// Injected Ignite.
+        /// </summary>
         [InstanceResource]
         public IIgnite Ignite { get; set; }
         
+        /** <inheritdoc /> */
         public bool Invoke(ICacheEntry<int, Foo> entry)
         {
+            if (entry.Key == FailKey)
+            {
+                throw new SecurityException("Crash in filter");
+            }
+            
             var cache = Ignite.GetCache<int, Foo>(CacheName);
+            var nearVal = cache.LocalPeek(entry.Key, CachePeekMode.Platform);
 
-            Foo _;
-            var hasNearVal = cache.TryLocalPeek(entry.Key, out _, CachePeekMode.Platform);
-
-            Assert.IsFalse(hasNearVal);
-            Assert.AreEqual(entry.Key, entry.Value.Bar);
+            Assert.AreSame(nearVal, entry.Value);
 
             return true;
         }
